Added GazeTracker and MouseDetector
Added a GazeTracker and MouseDetector. When running on Unity, these can be used, along with GestureManager, to subsititute mouse command for the HoloLens gaze and gesture features.
Also added a pausable trail to the Red Bouncing Ball. The force on the ball also varies as a function of how far the ball is from the central axis (x=0, z=0).